refactor(ocr): extract worker planning and parallel runner into engine/workers
